Extract anonymous chat logic
