+++ src/components/index.ts
@@ -1,2 +1,4 @@
 export { default as ChatUI } from "./ChatUI";
 export { default as MessageInput } from "./MessageInput";
+export { default as DataFastWidget } from "./DataFastWidget";
+export { default as MessageInputOverlay } from "./MessageInputOverlay";
