+++ src/hooks/index.ts
@@ -1,1 +1,2 @@
 export { useMessageInput } from "./useMessageInput";
+export { useChatUI } from "./useChatUI";
